requirements.txt
Django==4.2.7
django-crispy-forms==1.14.0
django-filter==23.3
Pillow==10.0.1
django-cleanup==8.0.0